* commit 'f1fa4679183a54d9dab0051abd2002d4a48203e6':
  chartType/visualizationType setting to dataset now done at openDataset, instead of chartType setter. Fixes chart type changing "on the fly"
  Cosmetics (changing the logError to be like the string exception logError).
  Fixing "warning C4101: 'ex': unreferenced local variable" compiler warning
  Fixing "warning C4244: '=': conversion from 'double' to 'int', possible loss of data" compiler warnings
  Fixing "warning C4018: '<': signed/unsigned mismatch" compiler warnings
  Fixing "anachronism used: qualifiers on reference are ignored" compiler warnings

+++ src/dmDataset.cpp
@@ -146,7 +146,7 @@ dmRasterImgData * dmDataset::getRasterData(
     coord &topLeftOut, coord &botRightOut,
     int &imgWidth, int &imgHeight)
 {
-    int n;
+    unsigned int n;
     int imgOffsetX, imgOffsetY;
 
     if (!_dstDataset)
@@ -350,13 +350,13 @@ std::pair<int, int> dmDataset::getRasterPixelOffsetAt(coord point, bool roundDow
 
     if (roundDownPixelOffset)
     {
-        pxOffsetX = std::floor(offsetX);
-        pxOffsetY = std::floor(offsetY);
+        pxOffsetX = (int)std::floor(offsetX);
+        pxOffsetY = (int)std::floor(offsetY);
     }
     else
     {
-        pxOffsetX = std::ceil(offsetX);
-        pxOffsetY = std::ceil(offsetY);
+        pxOffsetX = (int)std::ceil(offsetX);
+        pxOffsetY = (int)std::ceil(offsetY);
     }
 
     return std::pair<int, int>(pxOffsetX, pxOffsetY);
@@ -548,7 +548,7 @@ GDALDataset * dmDataset::visualizeDataset(GDALDataset *dsToVisualize)
     // put gdaldem processing flags into a c string array
     std::vector<std::string> optionsVec = getGdaldemOptionsVec();
     char** optionsArr = new char*[optionsVec.size() + 1];
-    for (int i = 0; i<optionsVec.size(); i++)
+    for (unsigned int i = 0; i<optionsVec.size(); i++)
     {
         optionsArr[i] = (char*)optionsVec[i].c_str();
     }
